feat : done updates based on new requirements

+++ charts-dashboard-template-backend/src/main/java/com/example/charts_dashboard_template_backend/gdp/model/CountryGdpData.java
@@ -23,7 +23,7 @@ public class CountryGdpData {
     private String country;
 
     @Column(nullable = false)
-    private int gdp;
+    private long gdp;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "gdp_data_id")
